Fixing scrolls of door/stair location and treasure detection
The previous implementation seemed to short circuit on the first condition of the `if` statement, causing the part after the `or` operator to be ignored. This resulted in `Scrolls of Door/Stair Location` to only show doors, and `Scrolls of Treasure Detection` to only show buried treasure.

Doors and stairs, as well as buried treasure and treasure, should properly show now.